refactoring : currency routes and test it

+++ app/Http/Controllers/Api/V1/Dashboard/CurrencyController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\V1\Dashboad\CurrencyRequest;
 use App\Http\Resources\Dashboard\CurrencyResource;
 use App\Models\Currency\Currency;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 
 class CurrencyController extends Controller
@@ -15,6 +16,17 @@ class CurrencyController extends Controller
     public function index(Request $request)
     {
         $currencies = Currency::latest()->paginate((int)($request->perPage ?? 10));
+
+        return CurrencyResource::collection($currencies)
+            ->additional([
+                'message' => 'success',
+                'status' => true
+            ]);
+    }
+
+    public function archive(Request $request)
+    {
+        $currencies = Currency::onlyTrashed()->latest()->paginate((int)($request->perPage ?? 10));
 
         return CurrencyResource::collection($currencies)
             ->additional([
@@ -75,11 +87,19 @@ class CurrencyController extends Controller
 
     public function destroy(Currency $currency)
     {
+        if ($currency->countries()->exists()) {
+            return response()->json([
+                'status' => false,
+                'message' =>  trans('dashboard.general.has_relationship_cannot_delete'),
+                'data' => null
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
         $currency->delete();
 
-        response()->json([
+        return CurrencyResource::make($currency)
+        ->additional([
             'status' => true,
-            'message' => trans("dashboard.general.has_relationship_cannot_delete"), 'data' => null
+            'message' =>  trans('dashboard.general.success_archive'),
         ]);
     }
 
@@ -87,9 +107,9 @@ class CurrencyController extends Controller
     public function restore($id)
     {
 
-        $currency = Currency::findOrFail($id)->restore();
+        $currency = Currency::onlyTrashed()->findOrFail($id);
 
-        // $currency->restore();
+        $currency->restore();
 
         return CurrencyResource::make($currency)
             ->additional([
@@ -100,16 +120,15 @@ class CurrencyController extends Controller
     }
 
 
-    public function forceDelete(Currency $currency)
+    public function forceDelete($id)
     {
-
-
+        $currency = Currency::onlyTrashed()->findOrFail($id);
         $currency->forceDelete();
 
         return CurrencyResource::make($currency)
             ->additional([
                 'status' => true,
-                'message' => trans('dashboard.general.success_delete')
+                'message' =>  trans('dashboard.general.success_delete'),
             ]);
     }
 }
